PASSWORD VALIDATION AND EXPOSURE
The password is allowed to be too short.  Changed the minimum
lenght to 7.  Also, when creating a new user, the password is
shown as the user types.  This could potentially lead to your
password being stolen by someone looking over your shoulder.
Changing the standard text field to a password field for the
password.