SESSION HIJACKING
The secret_key_base, used to incrypt the sesson is very short and
guessable.  Using rake secret, made a new secret key for development
and test.

In reality, this is not a vulnerability because the production
secret key base is an environment variable and as long as that
environment variable is sufficiently long and random, your sessions
will be properly encrypted.