SQL INJECTION and PRIVILEGE ESCALATION
The search feature string interpolates the search term into
a SQL query.  This allows an attacker to modify the search term.

If the attacker wants to find a post that includes keyword from
the set of private posts but does not have admin privileges, the
attacker could craft a search term like this:

KEYWORD%' OR content::TEXT like '%ANYTHING

This makes the AND public=true portion of the sql search term unimportant
and allows the attacker to search through private posts.  To fix this
attack, use prepared statements instead.